--- token-generator/cmd/main.go
@@ -1,23 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"log"
-	"os"
-
-	"github.com/GoogleCloudPlatform/functions-framework-go/funcframework"
-	generator "gruvee.com/tokengen"
-)
-
-func main() {
-	funcframework.RegisterHTTPFunction("/generateToken", generator.GenerateToken)
-	// Use PORT environment variable, or default to 8080.
-	port := "8080"
-	if envPort := os.Getenv("PORT"); envPort != "" {
-		port = envPort
-	}
-
-	if err := funcframework.Start(port); err != nil {
-		log.Fatalf("funcframework.Start: %v\n", err)
-	}
-}
