+++ tokengen/go.mod
@@ -1,0 +1,9 @@
+module github.com/PixelogicDev/Gruvee-Backend/tokengen
+
+require (
+	cloud.google.com/go/firestore v1.1.1 // indirect
+	firebase.google.com/go v3.12.0+incompatible
+	github.com/GoogleCloudPlatform/functions-framework-go v1.0.0
+	github.com/gogo/protobuf v1.2.0 // indirect
+	google.golang.org/api v0.17.0
+)
